Parser and listBox updates
The text file is now read and converted to a string. the string is
converted to a character array. the character array is read and creates
a list of resources. the list of resources is read and and 2 text boxes
are made for each resource, 1 for the name and 1 for all of the other
strings (this one has it's visibility set to collapsed)