Phone Calling Works for Directory
The only thing left is the aesthetic of having the info not run off the
screen. It SHOULD be a quick fix but 40 mins has proved unfruitfull (for
the record the phone calling ability took but 10 minutes to write and
test). I'll clean up the comments in my final push before I leave for
microsoft.